Python GUI/requirements.txt
pillow
python-socketio
mysql-connector-python
websocket-client
PyMuPDF
python-docx
PyPDF2 
firebase-admin 